+++ packages/cli/CHANGELOG.md
@@ -11,7 +11,7 @@
 - [#181](https://github.com/mesg-foundation/js-sdk/pull/181) Add util to get hash from a transaction log
 - [#186](https://github.com/mesg-foundation/js-sdk/pull/186) Create/delete process based on LCD endpoint
 - [#187](https://github.com/mesg-foundation/js-sdk/pull/187) Add `account:transfer` command
-- [#187](https://github.com/mesg-foundation/js-sdk/pull/187) Transfer token when create a process
+- [#187](https://github.com/mesg-foundation/js-sdk/pull/187) Transfer token when creating a process
 
 #### Bug fixes